Break the remaining isolated import cycles via leaf extraction
Resolves the last isolated strongly-connected components flagged by
lint:circular by extracting shared types into leaf modules — no runtime
behavior changes.

- runtime/http-router ↔ routes/http-adapter: the adapter imported
  HTTPRouteDefinition back from the router. Moved the HTTP route definition
  vocabulary (HTTPRouteDefinition, RouteContext, RouteParams, RouteQueryParam,
  RouteBodySchema, RouteAdditionalResponse) into runtime/http-router-types.ts;
  both the router and adapter now import from the leaf.

- monitoring/resource-sampler ↔ monitoring/stall-capture: stall-capture
  imported the ResourceSample type back from resource-sampler. Moved the
  sample vocabulary (ResourceSample + its sibling interfaces) into
  monitoring/resource-sample-types.ts; resource-sampler, stall-capture, and
  the monitoring route import from the leaf.

- cli/lib/install-from-github ↔ cli/lib/plugin-marketplace: plugin-marketplace
  imported the FetchLike type back from install-from-github. FetchLike is a
  generic injected-fetch shape used across seven cli/lib modules, so moved it
  into cli/lib/fetch-like.ts and repointed every importer — also gets it out
  of the 1.5k-line install-from-github module.

- plugins/defaults/memory/v2/harness retriever ↔ trace: mutual type import.
  Moved RetrievalCost into a harness-types leaf so trace no longer imports
  retriever. (retriever itself remains in the daemon core SCC via router.ts —
  an unrelated, larger cycle — but the retriever↔trace back-edge is gone.)

Isolated circular SCCs: 4 → 0. The only strongly-connected component left is
the daemon core; every other module is now cycle-free. All moved symbols are
imported directly from the leaves (no re-export shims).
