+++ assistant/src/cli/lib/__tests__/install-from-github.test.ts
@@ -22,8 +22,8 @@ import { tmpdir } from "node:os";
 import { dirname, join } from "node:path";
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
+import type { FetchLike } from "../fetch-like.js";
 import {
-  type FetchLike,
   type GitRunner,
   installPlugin,
   InvalidPluginNameError,
