+++ assistant/src/cli/lib/__tests__/plugin-details.test.ts
@@ -15,7 +15,7 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
-import type { FetchLike } from "../install-from-github.js";
+import type { FetchLike } from "../fetch-like.js";
 import {
   getPluginDetails,
   PluginDetailsNotFoundError,
