+++ assistant/src/cli/lib/__tests__/plugin-marketplace.test.ts
@@ -7,7 +7,7 @@
 
 import { describe, expect, test } from "bun:test";
 
-import type { FetchLike } from "../install-from-github.js";
+import type { FetchLike } from "../fetch-like.js";
 import {
   fetchMarketplaceEntries,
   type MarketplaceEntry,
