+++ assistant/src/cli/lib/__tests__/upgrade-plugin.test.ts
@@ -21,8 +21,8 @@ import { tmpdir } from "node:os";
 import { dirname, join } from "node:path";
 import { afterEach, beforeEach, describe, expect, test } from "bun:test";
 
+import type { FetchLike } from "../fetch-like.js";
 import {
-  type FetchLike,
   type GitRunner,
   PluginSourceUnavailableError,
 } from "../install-from-github.js";
