+++ assistant/src/cli/lib/diff-plugin.ts
@@ -40,9 +40,9 @@ import { createTwoFilesPatch } from "diff";
 
 import { PRESERVED_ENTRIES } from "../../plugins/plugin-tree-walk.js";
 import { getWorkspacePluginsDir } from "../../util/platform.js";
+import type { FetchLike } from "./fetch-like.js";
 import {
   DEFAULT_PLUGIN_REF,
-  type FetchLike,
   type GitRunner,
   materializePluginTree,
   type PluginFetchSource,
