+++ assistant/src/cli/lib/fetch-like.ts
@@ -1,0 +1,11 @@
+/**
+ * Minimal `fetch` shape used by the plugin CLI helpers.
+ *
+ * Narrower than `typeof fetch` because Bun's `fetch` carries a `preconnect`
+ * static these callers do not need — pinning to the wider type would force
+ * every caller to construct a fully-featured Bun fetch.
+ */
+export type FetchLike = (
+  input: RequestInfo | URL,
+  init?: RequestInit,
+) => Promise<Response>;
